Fix lab skeletons inconsistencies
* the .sln file inside lab06 was wrongly named lab07-win.sln
* the projects inside lab09 had inconsistent names; now all follow the
  schema: <task_number>-<task_name>
